Revert "revert type shuffling & DRY code out"
This reverts commit 9da09509e520a5ac80f28503fd648f03e3c8533b.

+++ packages/angel_core/src/messages/registrar.rs
@@ -1,4 +1,4 @@
-use cosmwasm_std::{Addr, Api, StdResult};
+use cosmwasm_std::{Addr, Api, Decimal, StdResult};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
@@ -9,7 +9,7 @@ pub struct MigrateMsg {}
 pub struct InstantiateMsg {
     pub accounts_code_id: Option<u64>,
     pub treasury: String,
-    pub tax_rate: u64,
+    pub tax_rate: Decimal,
     pub default_vault: Option<Addr>,
 }
 
@@ -51,7 +51,7 @@ pub struct UpdateConfigMsg {
     pub accounts_code_id: Option<u64>,
     pub index_fund_contract: Option<String>,
     pub treasury: Option<String>,
-    pub tax_rate: Option<u64>,
+    pub tax_rate: Option<Decimal>,
     pub approved_charities: Option<Vec<String>>,
     pub default_vault: Option<String>,
     pub guardians_multisig_addr: Option<String>,
